+++ pyzeebe/__init__.py
@@ -7,6 +7,6 @@ from pyzeebe.credentials.oauth_credentials import OAuthCredentials
 from pyzeebe.job.job import Job
 from pyzeebe.job.job_status import JobStatus
 from pyzeebe.task.exception_handler import ExceptionHandler
-from pyzeebe.task.task_decorator import TaskDecorator
+from pyzeebe.task.types import TaskDecorator
 from pyzeebe.worker.task_router import ZeebeTaskRouter
 from pyzeebe.worker.worker import ZeebeWorker
