--- pyzeebe/task/task_decorator.py
@@ -1,5 +1,0 @@
-from typing import Callable
-
-from pyzeebe.job.job import Job
-
-TaskDecorator = Callable[[Job], Job]
